Removed buildtime dependency on ed for bc
For some reason, the bc developers recently decided to start forcing a compile
time dependency on ed. This struck me as rather odd since no one really uses ed
anymore. This causes a problem, as ed is not available when bc is first built
in phase 2. Fortunately, Linux from Scratch has published a workaround. We have
applied that in the bc.buildinfo.