Moved c_rehash into cacertificates
This script (residing in /usr/bin) is required to build cacertifiates. It was provided by the OpenSSL package; however, that has been removed in Cucumber Linux 2.0 and an equivalent script is not provided by LibreSSL. We have added OpenSSL's version of the script to the cacertificates package to satisfy the build time dependency.