Upgraded bind-client to 9.11.4
net-base/bind-client upgraded from 9.11.3 to 9.11.4 to fix CVE-2018-5738, a
security vulnerability that allowed for unintended clients to perform recursive
queries, due to improper handling of the configuration. This vulnerability
likely affects only the bind-server package; however, this has not been
confirmed to be the case, so we have upgraded the client package as well just
to be safe. For more information see:
	https://security.cucumberlinux.com/security/details.php?id=467
	https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2018-5738
	https://kb.isc.org/article/AA-01634/0/BIND-9.11.4-Release-Notes.html

Also deleted the LibreSSL patches as equivalent patches have been applied
upstream in 9.11.4 and the old patches are therefore no longer necessary.

* SECURITY FIX *