Renamed phase 3 to phase 4
Originally, the build process was going to be split up into 3 phases as follows:
Phase 1: LFS Chapter 5; build a temporary /tools system to use when compiling
         actual system.
Phase 2: LFS Chapter 6 plus a few packages; build a basic system.
Phase 3: BLFS; build the remaining packages.

This created an issue for two reasons: one, the deliniation between phases 2
and 3 is not as clear as one may think. This made it difficult to decide which
packages were important enough to be considered part of the "basic" system.
Two, the ports tree would not become self hosted until the end of Phase 3. This
created an issue, as it effectively required you to build the entire system in
order for it to be able to recompile itself. In reality, only a small subset of
the phase 3 packages are actually necessary for this.

So in response, we have changed the build process to now use 4 phases as
follows:
Phase 1: Construct a temporary system in /tools that will be used to build the
         final system.
Phase 2: Build an initial bootstrap system. At the end of phase 2, the system
	 will be capable of running itself (without the /tools directory), but
	 will not be capable of compiling itself.
Phase 3: Build the necessary packages for the /cucumber directory of the ports
	 tree to become entirely self hosted. At the end of phase 3, the system
	 will be capable of both running *and* compiling itself; however, it
	 will not be complete yet.
Phase 4: Build the remaining packages in the /cucumber directory of the ports
	 tree that were not build in phases 2 or 3. At the end of phase 4, the
	 system will be complete.

This commit marks the first part of this change: changing the bulk of what was
previously Phase 3 to Phase 4.