Added mariadb and upgraded it to 10.3.9
Update Note:

There have been some important changes to the mariadb configuration file,
/etc/mysql/my.cnf. Namely, the option `innodb_additional_mem_pool_size` has
been removed in Mariadb 10.3. For Maraidb 10.1, this option was set by default
in my.cnf. This causes Mariadb to fail to start post upgrade unless the
offending line is removed.

Taking one of the following actions is recommended:
  1. Delete/comment out the line `innodb_additional_mem_pool_size = 2M` in
     /etc/mysql/my.cnf.
  2. Use the new my.cnf that is shipped with the updated package. This can be
     done by either choosing the 'REPLACE' new file action when upgrading the
     package or doing an `mv /etc/mysql/my.cnf /etc/mysql/my.cnf.old && mv
     /etc/mysql/my.cnf.new /etc/mysql/my.cnf`.

This note is included in the DRAFT Cucumber Linux 2.0 upgrade notes at:
https://cucumberlinux.com/~scott/cucumber-2.0/DRAFT_upgrade_notes