Changed portmake to use su instead of sudo
Previously, pormake used sudo run commands as the privilege separation user.
This causes issues in situations where you are trying to use portmake, but sudo
is not installed (such as when building Cucumber Linux from Scratch). To
rectify this issue, portmake has been changed to use `su -c` to drop privileges
instead.

Note that this requires portmake to be run as root now; `su -c` prompts for a
password otherwise. Also added a check for that.