Added metadata directory and refactored
Added a new metadata directory to hold metadata about the ports tree and the
distribution.

Also refactored the distribution version information to all reside in
metadata/distribution-release. This allows for the distribution release
information to be changed for the entire ports tree from a single file, which
will in turn make releasing new versions of the distribution easier.

Refactored the lsb-info package to use metadata/distribution-release.

Added note about the metadata directory to the README.