Updated distribution-release
Removed the .alpha.1 suffix from the distribution version number. Having
this suffix has caused unintended consequences, such as necessitating a
new branch be created on the mirror servers whenever a new alpha/beta/rc
version is released. This will enable us to use a single branch for the
entirety of a Cucumber Linux version's development and release cycle.