lsb-info: updated version determination mechanism
The distribution-release file no longer contains .alpha/.beta/.rc at the
end of the DISTRIB_RELEASE variable. In order to keep it clear when a
development version of the package is installed, add ".$DISTRIB_TYPE" to
the end of the package version when $DISTRIB_TYPE is not "stable". When
it is "stable", then this is the stable release of that version of
Cucumber Linux, so adding the suffix is not necessary.