thunderbird: upgraded to 60.4.0
This update contains several bug fixes. It likely also contains security fixes;
however, Mozilla does not like to publicly disclose details about security
updates (or even indicate that an update is a security update) until a couple
of weeks after the update it released, so this cannot be confirmed at this
time. We have upgraded to be safe. For more information see:
	https://security.cucumberlinux.com/security/details.php?id=647
	https://www.thunderbird.net/en-US/thunderbird/60.4.0/releasenotes/

* SECURITY FIX *