portmake: added build number autoincremnatation
Portmake now has the ability to automatically increment the build number when
rebuilding a package, toggleable via the AUTOINCREMENT_BUILD_NUMBER variable in
portmake.conf. If this variable is set to true (which is now the default), the
build number will be set to the current build number plus 1 if the full package
name of the package being built is otherwise the same as the version already
installed. Setting the BUILD environment variable supersedes this.