portmake: added old package retirement support
Added the ability to "retire" "old packages" to portmake. When enabled, this
causes portmake to move the binary packages for all the previous builds of the
package out of the PACKAGE_DIR (/opt/packages by default) and to their
corresponding location in the OLD_PACKAGE_DIR (/opt/old_packages by default)
when the package is rebuilt.  For example, if links is being rebuilt (build 2),
then the file /opt/packages/cucumber/apps/links-2.13-x86_64-sysbuild18002_1.txz
will be moved to /opt/old_packages/cucumber/apps/. Additionally, any signature
files (.asc, .sig or an other extension on the package file name) will be moved
along with the package file.

This feature can be toggled via the MOVE_OLD_PACKAGES variable in portmake.conf.

The directory that old packages are moved to can be changed via the
OLD_PACKAGE_DIR variable in portmake.conf.