portutils: changed utilities from symlink to dir
Changed the utilties subdirectory in portutils from a symlink to a directory.
The contents of ports/utilties that are needed for the portutils package are
now symlinked directly into that directory. This prevents portmake from
attempting to update the portutils package every time there is a change
anywhere in the ports/utilties directory, even if that change does not effect
the portutils package.